Added Code Conventions file
This was for a project I was programming lead for. I used it both as a living example (perhaps a living guildpact) and a rules document. It also demonstrates the rough stylings of my personal code, and how I tend to set it up in terms of navigability and readability